requirements.txt
streamlit
opencv-python
ultralytics
streamlit-webrtc
numpy
av
pyserial
